# packages/contracts/generated/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AddEdgeBody": {
        "properties": {
          "confirmation": {
            "default": "confirmed",
            "title": "Confirmation",
            "type": "string"
          },
          "from_node_id": {
            "format": "uuid",
            "title": "From Node Id",
            "type": "string"
          },
          "to_node_id": {
            "format": "uuid",
            "title": "To Node Id",
            "type": "string"
          }
        },
        "required": [
          "from_node_id",
          "to_node_id"
        ],
        "title": "AddEdgeBody",
        "type": "object"
      },
      "AddNodeBody": {
        "properties": {
          "confirmation": {
            "default": "suggested",
            "title": "Confirmation",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "title": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "title"
        ],
        "title": "AddNodeBody",
        "type": "object"
      },
      "AppendEventBody": {
        "properties": {
          "answer_revealed": {
            "default": false,
            "title": "Answer Revealed",
            "type": "boolean"
          },
          "assessment_item_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Item Id"
          },
          "assessment_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Revision Id"
          },
          "attempt_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Id"
          },
          "client_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Event Id"
          },
          "concept_node_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Node Id"
          },
          "course_id": {
            "format": "uuid",
            "title": "Course Id",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "response_before_feedback": {
            "default": true,
            "title": "Response Before Feedback",
            "type": "boolean"
          },
          "substantive_hint": {
            "default": false,
            "title": "Substantive Hint",
            "type": "boolean"
          }
        },
        "required": [
          "course_id",
          "event_type"
        ],
        "title": "AppendEventBody",
        "type": "object"
      },
...
      "CreateAssessmentBody": {
        "properties": {
          "concept_node_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Node Id"
          },
          "course_id": {
            "format": "uuid",
            "title": "Course Id",
            "type": "string"
          },
          "item_type": {
            "title": "Item Type",
            "type": "string"
          },
          "prompt": {
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "sealed_answer": {
            "additionalProperties": true,
            "title": "Sealed Answer",
            "type": "object"
          },
          "sealed_rubric": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sealed Rubric"
          },
          "source_locators": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Locators"
          },
          "title": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "course_id",
          "title",
          "item_type",
          "prompt",
          "sealed_answer"
        ],
        "title": "CreateAssessmentBody",
        "type": "object"
      },
...
      "CreateCourseBody": {
        "properties": {
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          },
          "title": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "space_id",
          "title"
        ],
        "title": "CreateCourseBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RebuildBody": {
        "properties": {
          "course_id": {
            "format": "uuid",
            "title": "Course Id",
            "type": "string"
          }
        },
        "required": [
          "course_id"
        ],
        "title": "RebuildBody",
        "type": "object"
      },
...
      "RevealAttemptBody": {
        "properties": {
          "answer": {
            "default": false,
            "title": "Answer",
            "type": "boolean"
          },
          "attempt_id": {
            "format": "uuid",
            "title": "Attempt Id",
            "type": "string"
          },
          "hint": {
            "default": false,
            "title": "Hint",
            "type": "boolean"
          }
        },
        "required": [
          "attempt_id"
        ],
        "title": "RevealAttemptBody",
        "type": "object"
      },
...
      "StartAttemptBody": {
        "properties": {
          "assessment_item_id": {
            "format": "uuid",
            "title": "Assessment Item Id",
            "type": "string"
          },
          "assessment_revision_id": {
            "format": "uuid",
            "title": "Assessment Revision Id",
            "type": "string"
          },
          "client_attempt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Attempt Id"
          },
          "course_id": {
            "format": "uuid",
            "title": "Course Id",
            "type": "string"
          }
        },
        "required": [
          "course_id",
          "assessment_item_id",
          "assessment_revision_id"
        ],
        "title": "StartAttemptBody",
        "type": "object"
      },
      "SubmitAttemptBody": {
        "properties": {
          "answer_revealed": {
            "default": false,
            "title": "Answer Revealed",
            "type": "boolean"
          },
          "assessment_item_id": {
            "format": "uuid",
            "title": "Assessment Item Id",
            "type": "string"
          },
          "assessment_revision_id": {
            "format": "uuid",
            "title": "Assessment Revision Id",
            "type": "string"
          },
          "client_attempt_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Client Attempt Id",
            "type": "string"
          },
          "confidence": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Confidence",
            "type": "string"
          },
          "course_id": {
            "format": "uuid",
            "title": "Course Id",
            "type": "string"
          },
          "hint_revealed": {
            "default": false,
            "title": "Hint Revealed",
            "type": "boolean"
          },
          "response": {
            "additionalProperties": true,
            "title": "Response",
            "type": "object"
          }
        },
        "required": [
          "course_id",
          "assessment_item_id",
          "assessment_revision_id",
          "response",
          "confidence",
          "client_attempt_id"
        ],
        "title": "SubmitAttemptBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Canonical domain API. Adds documents, memory, context, sources, jobs, and ingestion routes.",
    "title": "Memdot Core API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/learning/assessments": {
      "post": {
        "operationId": "create_assessment_api_v1_learning_assessments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssessmentBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Assessment",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/assessments/{assessment_item_id}/revisions/{revision_id}/attempt": {
      "get": {
        "operationId": "get_attempt_view_api_v1_learning_assessments__assessment_item_id__revisions__revision_id__attempt_get",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assessment Item Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Attempt View",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/attempts": {
      "post": {
        "operationId": "submit_attempt_api_v1_learning_attempts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAttemptBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Attempt",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/attempts/reveal": {
      "post": {
        "operationId": "reveal_attempt_api_v1_learning_attempts_reveal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevealAttemptBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reveal Attempt",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/attempts/start": {
      "post": {
        "operationId": "start_attempt_api_v1_learning_attempts_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAttemptBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Attempt",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/courses": {
      "post": {
        "operationId": "create_course_api_v1_learning_courses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Course",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/courses/{course_id}/edges": {
      "post": {
        "operationId": "add_edge_api_v1_learning_courses__course_id__edges_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEdgeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Edge",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/courses/{course_id}/nodes": {
      "post": {
        "operationId": "add_node_api_v1_learning_courses__course_id__nodes_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNodeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Node",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/events": {
      "post": {
        "description": "Public arbitrary learner-event append is removed (Round 2).\n\nEvidence must flow through the server-owned attempt lifecycle only.",
        "operationId": "append_event_api_v1_learning_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendEventBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append Event",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/evidence/rebuild": {
      "post": {
        "operationId": "rebuild_evidence_api_v1_learning_evidence_rebuild_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RebuildBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rebuild Evidence",
        "tags": [
          "learning"
        ]
      }
    },
...
  }
}